renamed email reminder construct filename
